Size Animation and Audio time
The size of the animation now is correctly handled by the LineInstance:
The recorder records all the keyframes, which means that the line reacts to the music during recording time. Now the property of the size is removed by the AnimationsTimeLine script.

When recording, the audio time starts where the animation will be recorded