Changed data structure and more
the new timeLine is a List<Clip>(), where Clip is a struct with a float time and string name. This makes it way easier to work with it.

Now Only three components are required to start animating a Line. These are independent from each other, so you can delete the last two after you are done animating:
1. Animate: Animates the line
2. ClipRecorder: Allows to record the line at runtime
3.TimeLine: Allows to modify the timeLine in Unity Edit Mode (not run time)

The Animate Script still requires the prefab to be instantiated that requires the LineInstance script. Probably this will change in the future.

Added new functionalities to timeLine:
-Modify Time
-Offset Times
-Copy Animation