Debugging tools
Added to Package Manager:
-Lunar Console
-Runtime Inspector

To use, in Prefabs, grab the Debug Object and drag it into the scene.
When running an app on mobile, you will see the Hierarchy and Inspector on the top of the screen. To enable the console, swipe with two finger in a up-down motion. If the Console occupies all the screen, you can find an option to resize the windows (I like it on the bottom)